refactor: pages folder structure
