+++ src/Router/Router.tsx
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
 import { Route, Routes } from 'react-router-dom';
 
-import HomePage from '../pages/HomePage';
-import AboutPage from '../pages/AboutPage';
-import NotFoundPage from '../pages/NotFoundPage';
+import { HomePage, AboutPage, NotFoundPage } from '../pages';
 
 class Router extends Component {
   render() {
