+++ src/pages/index.ts
@@ -1,0 +1,5 @@
+import HomePage from './HomePage/HomePage';
+import AboutPage from './AboutPage/AboutPage';
+import NotFoundPage from './NotFoundPage/NotFoundPage';
+
+export { HomePage, AboutPage, NotFoundPage };
